refactor: 코드리뷰 반영해서 모달 관련 작업은 Header에서 하도록 변경
